pyproject.toml
[tool.poetry]
name = "handwritten_ocr"
version = "0.1.0"
description = ""
authors = ["adimyth <mishraaditya6991@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
tensorflow = "2.5.0"
opencv-python = "^4.5.2"
python-Levenshtein = "^0.12.2"
sklearn = "^0.0"
pandas = "1.1.5"
matplotlib = "^3.4.2"
seaborn = "^0.11.1"
black = "^21.5b1"
isort = "^5.8.0"
tqdm = "^4.60.0"
attrdict = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
